% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/maglite-package.r
\docType{package}
\name{maglite}
\alias{\%$\%}
\alias{\%<>\%}
\alias{\%>\%}
\alias{\%T>\%}
\alias{maglite}
\alias{maglite-package}
\title{maglite.}
\description{
Makes available just the pipe operators from \pkg{magrittr} (hence the name,
\strong{mag}rittr \strong{lite}). This is useful if you want to avoid any
name conflicts that the magrittr aliases introduce.
}
\details{
maglite makes the following functions available:
  \itemize{
     \item pipe \code{\link[magrittr]{\%>\%}}
     \item compound assignment pipe \code{\link[magrittr]{\%<>\%}}
     \item exposition pipe \code{\link[magrittr]{\%$\%}}
     \item tee pipe \code{\link[magrittr]{\%T>\%}}
  }
}
\seealso{
\link[magrittr]{magrittr}
\href{https://cran.r-project.org/web/packages/magrittr/vignettes/magrittr.html}{magrittr-vignette}
}

